FIX: windows 10 build with msvc 2019 part two FIX: compile warnings

+++ code/renderer/tr_glshader.cc
@@ -561,7 +561,7 @@ void GLShaderManager::buildAll( )
 	while ( !_shaderBuildQueue.empty() )
 	{
 		GLShader *shader = _shaderBuildQueue.front();
-		size_t numPermutations = 1 << shader->GetNumOfCompiledMacros();
+		size_t numPermutations = size_t(1) << shader->GetNumOfCompiledMacros();
 		size_t i;
 
 		for( i = 0; i < numPermutations; i++ )
@@ -582,7 +582,7 @@ void GLShaderManager::buildAll( )
 
 void GLShaderManager::InitShader( GLShader *shader )
 {
-	shader->_shaderPrograms = std::vector<shaderProgram_t>( 1 << shader->_compileMacros.size() );
+	shader->_shaderPrograms = std::vector<shaderProgram_t>( size_t(1) << shader->_compileMacros.size() );
 	memset( &shader->_shaderPrograms[ 0 ], 0, shader->_shaderPrograms.size() * sizeof( shaderProgram_t ) );
 
 	shader->_uniformStorageSize = 0;
